pyproject.toml
[tool.poetry]
name = "autoupdate"
version = "0.1.2"
description = ""
authors = ["vol4ok <vol4oks.d@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"

[tool.poetry.scripts]
main = "autoupdate.__main__:main"
autoupdate = "autoupdate:main"
test = "tests.start:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
